llm/requirements.txt
httpx==0.27.0
langchain==0.1.10
python-dotenv==1.0.1
